libexpr/lexer: directly parse negative ints as negative
Breaks parsing of things like: `0 -100`, `12-123` etc.

This change makes the sign for negative integers part of the number
instead of considering it as the negation operator which would negate
the parsed positive integer when evaluating.

The intent of this change is to fix the issue that not all integers
representable in nix could actually be written as an integer literal:
`-9223372036854775808`, the smallest integer representable by nix
currently causes a parse error if written like that, but can still be
obtained by writing `-9223372036854775807 - 1`. The error (`invalid
integer '9223372036854775808'`) reveals the issue: The integer is parsed
as positive and negated after parsing, but the biggest representable
integer in nix, `9223372036854775807`, is actually one to small for the
job.

The change has the following side effects:

* There is now a distinction between proper negative integer literals
  and negated, positive integers: `-12368` is parsed as a negative
  integer, whereas `- 12368` and `-(12368)` are parsed as negation
  operators followed by a positive integer. This means, parsing
  `-(9223372036854775808)` still fails, but this is probably more
  intuitive.
* We introduce the backwards incompatibility that older Nix version
  won't parse `-9223372036854775808`, but newer ones will.

This change has mainly tested by throwing `nix-instantiate --parse`
against existing nix source files:

* I parsed all proper `.nix` files in nixpkgs with nix stable and fixed
  the remaining parse errors in this PR:
  https://github.com/NixOS/nixpkgs/pull/114290

  And then parsed them again with the patched version of to make sure
  I'd get the same result (no parse failures).

  ```
  find ~/src/nix/nixpkgs/ -name '*.nix' \
    -and -not -path '**/python/cpython/docs/template.nix' \
    -and -not -path '**/nixpkgs/.git/**' -print \
    | while read line; do ./outputs/out/bin/nix-instantiate --parse "$line" > /dev/null; done
  ```

  I've also done similar checks against other projects using nix.

* Ran nix unit tests and lang tests, the latter of which reveal
  unforeseen consequences.